requirements_dev.txt
-r requirements.txt

factory_boy==2.3.1
fake-factory

django-debug-toolbar==1.2
django-extensions
